Cleaned up isEventHandled
